fix: slice client usecase into multiple files
Separate usecase/client.go into multiple files according to method type
- client.go
+ create_client
+ update_client
+ get_clients
